requirements.txt
# requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
python-telegram-bot>=20.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0
asyncio
smtplib
email-mime-text
logging
datetime
typing
hashlib
random